backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pandas
scikit-learn
timescale
yfinance
transformers
tqdm
python-dotenv
textblob
requests
nltk
dash>=2.8.0
dash-bootstrap-components>=1.0.0
plotly==5.17.0
xgboost
prophet
tensorflow
numpy>=1.20.0,<2.0.0
ta
mplfinance
schedule
httpx
aiofiles
asyncpg
aiohttp
openai>=1.0.0